requirements.txt
asgiref==3.9.1
Django==5.2.6
django-mathfilters==1.0.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
